+++ source/texact.py
@@ -13,6 +13,7 @@ from reviewers.reviewer_figure import Reviewer_Figure
 from reviewers.reviewer_inthis import Reviewer_Inthis
 from reviewers.reviewer_math import Reviewer_Math
 from reviewers.reviewer_reflabel import Reviewer_RefLabel
+from reviewers.reviewer_section import Reviewer_Section
 from reviewers.reviewer_unsure import Reviewer_Unsure
 from template_check import get_template
 
@@ -229,6 +230,7 @@ def main():
         reviewers = [
             Reviewer_Inthis(printer),
             Reviewer_RefLabel(printer),
+            Reviewer_Section(printer),
             Reviewer_Math(printer, template),
             Reviewer_Casing(printer, config.lint.casing),
             Reviewer_Figure(printer, file_path),